chore(passport): Fix passport routes and create app file to setup
